SidePanel: remove unnecessary binding
The purpose of the swarmCurrentConversationList in SidePanel is to allow
the user to select the initial list of members when creating a group
conversation. The binding on currentIndex is not necessary for this and
has been causing a segfault in the WelcomePage test since the upgrade
from Qt 6.6 to Qt 6.8.

GitLab: #2108
Change-Id: Ie9341f114ba2a8abc907081e9041b85a4f260333